pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = "FastAPI LLM application"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "gunicorn>=21.2.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0",
    "google-generativeai>=0.5.0",
    "python-multipart>=0.0.6",
]